Create new file to avoid import cycles
