+++ src/python/pants/backend/awslambda/python/rules_test.py
@@ -13,6 +13,7 @@ from zipfile import ZipFile
 
 import pytest
 
+from pants.backend.awslambda.python.aws_architecture import AWSLambdaArchitecture
 from pants.backend.awslambda.python.rules import (
     PythonAwsLambdaFieldSet,
     PythonAwsLambdaLayerFieldSet,
@@ -21,11 +22,7 @@ from pants.backend.awslambda.python.rules import (
     package_python_aws_lambda_layer,
 )
 from pants.backend.awslambda.python.rules import rules as awslambda_python_rules
-from pants.backend.awslambda.python.target_types import (
-    AWSLambdaArchitecture,
-    PythonAWSLambda,
-    PythonAWSLambdaLayer,
-)
+from pants.backend.awslambda.python.target_types import PythonAWSLambda, PythonAWSLambdaLayer
 from pants.backend.awslambda.python.target_types import rules as target_rules
 from pants.backend.python.goals import package_pex_binary
 from pants.backend.python.goals.package_pex_binary import PexBinaryFieldSet
